dropbox-mongodb/backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB
pymongo==4.6.0
motor==3.3.2

# Azure services
azure-servicebus==7.11.4
azure-storage-blob==12.19.0

# Dropbox
dropbox==11.36.2

# OpenAI for embeddings
openai==1.54.3

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Additional utilities
python-jose[cryptography]==3.3.0

